cli: Do a dry run before executing rsync --delete
Alter the deploy command to first do a dry run check with rsync listing
files to be deleted, and to ask for confirmation if it will delete any
files. This prevents an issue where it deletes all your files if you
mess up your config, or at least makes it more your fault if you still
press 'yes'.